PROPERTIES Ask the engine what holds for inputs nobody wrote down
The suite checked the engine against scenes someone had written out: a ball
dropped three metres, a stack of five boxes, a wheel on a slope. Those say the
physics is right for the cases we thought of. `tests/test_properties_*.py` ask a
different question -- whether the invariants survive inputs nobody thought of --
over shapes an asset pipeline produces and a level never means to contain: a
collider with no radius, a box flattened to a plane, a "triangle" whose three
vertices are one point, a dozen bodies already inside one another, colliders
spanning a millimetre to a kilometre, a hundred-kilogramme crate resting on a
hundred-gramme one. Hypothesis generates them, shrinks a failure to the smallest
case that still fails, and remembers it.

It found three defects.

**A triangle with no area produced a contact with no normal.** All three
capsule-triangle implementations -- scalar, numpy batch and the accelerator --
reported a hit off a zero-area triangle with a normal of (0, 0, 0), which the
character controller then multiplies by. A zero-area triangle is not a surface:
it has no side to be on, no direction to be pushed along, and whatever solid it
belongs to is bounded by its neighbours. It now makes no contact, against a box
as well as a capsule. The named case for it was already in the suite with the
right title and an assertion that let it through.

**The position pass over-corrected by the size of the manifold, and diverged.**
Each of a manifold's four points applied the full correction against the depth
the narrow phase measured, so a box resting on a floor was pushed four times as
far as it was in, came back further out of place than it started, and grew from
there. A 75kg box resting on a 1kg box drove the light one through the floor and
it fell for as long as the scene ran. Each contact is now corrected against what
the pass has already applied, and `SequentialImpulseSolver.max_correction`
(0.2m, per step) bounds what one step may do -- so a body that arrives deep
inside another climbs out over a few frames rather than being flung.

**`quat_to_axis_angle` returned an axis that was not a direction.** Recovering
it through `sqrt(1 - w*w)` cancels as `w` approaches one, which is where a body
turning slowly spends every frame; at 1e-7 radians the axis was a percent short.
It comes from the quaternion's own vector part now, and the angle from
`arctan2`. `threaded.py` publishes this to whatever is drawing the world.

Two other answers turned out to be ambiguous rather than wrong, and are written
down instead of changed: coincident colliders and pairs touching at exactly no
depth pick a way out rather than computing one, and a clipped box manifold
reports each point's own penetration rather than the distance that separates the
pair. Where the mass ratios reach is documented too -- past about a hundred to
one, several bodies deep, sequential impulses cannot pass the load down the
stack and the light body is squeezed out sideways.

Hypothesis is a CPython-only test dependency: it publishes no wheel for the PyPy
in the matrix. The property files skip where it is absent, so PyPy runs
everything else as before.

Lint and types become gates in the same piece of work, since they were declared
in the `dev` extra and run nowhere. Both read their whole configuration from
pyproject.toml, so `ruff check` and `mypy` in the checkout are the runs tox and
CI make; `tox -e lint` and `tox -e typecheck` are on the 3.12 row. Clearing them
is most of the diff: `typing.List` and `Optional[X]` give way to the builtin and
`|` forms the floor of 3.10 has had all along, imports are sorted, `zip` says
whether it means the lengths to match, and four parameters that nothing read are
gone from the signatures that claimed them. `ruff format` is configured but is
not a gate -- the array expressions are laid out so the arithmetic keeps its
shape, and the formatter sets them as filled paragraphs of code.

Verified on both accelerator paths, and against the projects built on this one:
twig-bb, glisteel and OpenGLContext all pass their physics tests unchanged.

+++ tests/test_character_impulse.py
@@ -3,8 +3,8 @@ import numpy as np
 import pytest
 
 from omi_physics import model
+from omi_physics.character import CharacterCapabilities, CharacterController
 from omi_physics.world import PhysicsWorld
-from omi_physics.character import CharacterController, CharacterCapabilities
 
 DT = 1.0 / 60.0
 GRAVITY = 9.81
@@ -30,7 +30,7 @@ def standing_character(caps=None):
 def fly_for(character, seconds):
     """Step ``seconds`` of simulation, returning the highest base height seen."""
     peak = character.base()[1]
-    for _ in range(int(round(seconds / DT))):
+    for _ in range(round(seconds / DT)):
         character.update(DT)
         peak = max(peak, character.base()[1])
     return peak
